CAEPSessionRevoked extends SSEvent, improve assertion test
I finally figured out the inheritance in the subclasses
to let Builder() work as desired extending SSEvent.

Turns out JSONObject does implement equals correctly, so we can simply assertEquals on two JSONObjects rather than convert them to text and compare the text.  Comparing text is fragile when the order of the claims in the output text does not have to have to be consistent, yet the objects are the same.