Genericize SSEvent
We have a problem with the Builder pattern used:
https://javadevcentral.com/effective-java-builder-pattern

that arises because we're using inheritance, but Foo.Builder().chain().of().additions().build() must go from deepest subclass methods to highest superclass methods, and must return the top level superclass object type, which will then make it hard to do any other reasoning about it.

As suggested here:
http://egalluzzo.blogspot.com/2010/06/using-inheritance-with-fluent.html

we employ Generics and abstract classes at the superclass level in SSEvent, which lets us return the subclass object in build().

We added an eventType member, so we also need to add setters and getters, equals() and hashCode();