CAEP Event Timestamps are required once again
As a result of the 03/30/2021 meeting, at which Credential Compromised
events moved into the RISC spec, nothing is forcing event_timestamp to
be optional any longer.  This should allow the change proposed the
previous week to make event_timestamp optional to be dropped, so it's
required once again.

Revert validator change to throw an exception if event_timestamp is not
present.

Signed-off-by: Matt Domsch <matt.domsch@sailpoint.com>