Add SEToken.parse(), NonstandardSSEvent
We want to be able to parse an incoming JSON string into a fully fleshed
JWTClaimsSet and event objects of our specific defined classes,
rather than leaving them all JSONObjects.  This lets us run
validate() on objects of the defined classes.

Minor changes to SSEvent and SSEventTypes to enable the parser.

One additional test with a nonstandard event name URL, and changes to
another to invoke the parser.

Signed-off-by: Matt Domsch <matt.domsch@sailpoint.com>