Improve parsing JSONObjects into specific SubjectIdentifiers
Now that we have subclasses of SubjectIdentifier for
specific subject formats with specific validation rules,
we need to parse JSONObjects into these specific
subclasses.

I had code in SEToken to do this, but that wasn't the
right place - it should be in SubjectIdentifier.  It
also didn't know about these new subclasses.

Move this code out of SEToken into SubjectIdentifier
and teach it how to use the new subclasses.

Signed-off-by: Matt Domsch <matt.domsch@sailpoint.com>